refactor(cloister): extract specialists registry seam (pan-2232)
